docs(v5): Added v5 docs to apib - Categories, whats new, translations, auth

+++ v5/Http/Controllers/CategoryController.php
@@ -201,6 +201,9 @@ class CategoryController extends V4Controller
     public function delete(int $id, Request $request)
     {
         $category = Category::withoutGlobalScopes()->find($id);
+        if (!$category) {
+            return self::make404();
+        }
         $this->authorize('delete', $category);
         $success = DB::transaction(function () use ($id, $request, $category) {
             $category->translations()->delete();
